pyproject.toml
[project]
name = "stproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "streamlit-cookies-manager>=0.2.0",
]
